+++ tests/basic_integration_test.rs
@@ -1,6 +1,7 @@
 use uk_areacodes::api;
 use uk_areacodes::data_source::json::UK;
 
+//noinspection SpellCheckingInspection
 #[test]
 fn it_returns_matches() {
     let data: Vec<api::Place> = serde_json::from_str(&UK).expect("JSON was not well-formatted");
@@ -14,6 +15,6 @@ fn it_returns_matches() {
     }
 
     if let Some(n) = api::binary_search(&data, 0, data.len() - 1, "01503") {
-        assert_eq!("Looe", data[n as usize].area);
+        assert_eq!("Looe", data[n].area);
     }
 }
